[tpmeventlog] Move Replay from `pcr` to `tpmeventlog`
