+++ pkg/tpmeventlog/errors.go
@@ -2,6 +2,8 @@ package tpmeventlog
 
 import (
 	"fmt"
+
+	pcr "github.com/9elements/converged-security-suite/v2/pkg/pcr/types"
 )
 
 // ErrRead means unable to read from the io.Reader
@@ -65,4 +67,30 @@ type ErrNotSupportedHashAlgo struct {
 // Error implements interface `error`.
 func (err ErrNotSupportedHashAlgo) Error() string {
 	return fmt.Sprintf("not supported hash algorithm: 0x%x", err.TPMAlgo)
+}
+
+// ErrUnexpectedEventType means there was received an unexpected event type
+type ErrUnexpectedEventType struct {
+	Event  Event
+	Reason string
+}
+
+// Error implements interface `error`.
+func (err ErrUnexpectedEventType) Error() string {
+	return fmt.Sprintf("unexpected event type, reason: '%s'; event: '%#+v'", err.Reason, err.Event)
+}
+
+// ErrNotSupportedIndex means selected PCR index is not supported (yet?)
+type ErrNotSupportedIndex struct {
+	Index       pcr.ID
+	Description string
+}
+
+// Error implements interface `error`.
+func (err ErrNotSupportedIndex) Error() string {
+	result := fmt.Sprintf("PCR index %d is not supported", err.Index)
+	if len(err.Description) > 0 {
+		result += ": " + err.Description
+	}
+	return result
 }
